requirements.txt
anyio==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
h11==0.12.0
httpcore==0.14.7
httpx==0.22.0
idna==3.3
notion-client==0.9.0
python-dotenv==0.20.0
requests==2.27.1
rfc3986==1.5.0
sniffio==1.2.0
urllib3==1.26.9
